PVF_System_Framework/requirements.txt
requests>=2.28.0
httpx>=0.23.0
schedule>=1.2.0
psutil>=5.9.0
python-dotenv>=1.0.0
cryptography>=3.4.8
bcrypt>=4.0.0
PyJWT>=2.6.0
locust>=2.14.0
memory-profiler>=0.60.0
prometheus-client>=0.16.0
colorama>=0.4.6
click>=8.0.0
flask>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
bandit>=1.7.0
safety>=2.3.0
email-validator>=1.3.0
